Fix bug in zsh config
Zsh would output strange character sequences before lines when using
percent signs, as well as with some other characters. The function that
sets the window title to the current command was adding the full command
to the string it was pringing with prompt expantion. This caused issues
with lines with characters affected by prompt expantion, resulting in
some characters being printed on the line.

The print statement was split in two with prompt expantion not set for
the one containing the command string.