Rework config management
This commit changes how config files are managed. Previously they were
just using GNU Stow. This required a directory structure that was
annoying to deal with, and forced the config files to be regular
files. It also used symlinks which caused some confusion.

Now all files are organized in each program's directory how they make
sense, and an install script is used to add copies to the proper place.
This also allows the config files to be scripts whose output is to be
copied to where it needs to go, which opens the way for future changes.